- Update to 1.8.2 - Fix (hopefully) the IRQ reporting on 5.x, before, xosview was simply exiting/crashing
PR:		79338
Submitted by:	Ulrich Spoerlein <q AT uni dot de>

+++ sysutils/xosview/files/patch-swapmeter.cc
@@ -2,6 +2,6 @@
 +++ bsd/swapmeter.cc	Tue Aug 21 10:53:02 2001
 @@ -76,3 +76,3 @@
  void SwapMeter::getswapinfo( void ){
--  int total_int, free_int;
+-  unsigned long long total_int, free_int;
 +  int64_t total_int, free_int;
